% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{MicroBias.fit}
\alias{MicroBias.fit}
\title{Fit the log-linear model for bias factors}
\usage{
MicroBias.fit(
  otu.tab,
  x,
  offset = NULL,
  delta = NULL,
  cols.are.otus = TRUE,
  tol = 10^-8
)
}
\arguments{
\item{otu.tab}{Feature table of the observed abundances.Each row is one sample and each column is one taxon. It doesn't have to be normalized. However, data will be transformed into relative abundances for analysis.}

\item{x}{The covariate matrix that you want to test for bias.}

\item{offset}{The true relative abundances matrix that you want to test as an offset. It should be of the same dimension of the otu.tab. If NULL, it is assumed that all taxa that are present in each sample have the same abundance levels within the sample.}

\item{delta}{The indicator matrix of whether the corresponding taxon is present in each sample. Should be of the same dimension as otu.tab. If NULL, it is taken as all taxa that have nonzero reads in a sample is designed to be present in the sample.}

\item{tol}{Tolerance level for matrix inversion. Default is 10^-8. Should be a small number.}
}
\value{
coefficient estimate of the log linear model (beta.matrix); and some other variables that need to be used for inference.
}
\description{
This function fits the log-linear model to the data using the least squares.
}
